Add Module building support to the knowledge construction engine
I updated the LLVMModule class so that it has a inline-asm slot as well.

I also removed the ability to select loop or regions. In the current design you
get them both even if you don't want them.

It is up to the expert system to handle the encapsulation of instances into the
proper modules.