Modified the compiler class
I added a reference to an LLVMContext object that will be provided to the
compiler object before compilation occurs. This LLVMContext manages the global
"llvm" data that keeps track of the different data structures.

I also removed or modified the methods that explicitly take in a
CLIPSEnvironment object. Those entries came from an earlier idea that no longer
appleis now that we setup the compiler before execution.

I also added getters and setters for the CLIPSEnvironment and LLVMContext
objects.