Modified CLIPSEnvironment to be far more intelligent.
Instead of always creating a new environment when instanced, CLIPSEnvironment
now will only manage a given environment if no arguments are provided to the
constructor. A void pointer of an already preexisting environment will prevent
the destructor from deleting the environment.

This is necessary for Compilers that may have temporary environments as part of
a call from clips.