Implemented CompilerManager.cpp and made some major changes
First off, I made it so that if compile is called from CLIPS then the provided
environment is used instead of a hardcoded one. I also modified the compile and
interpret methods of both Compiler and CompilerManager to be far simpler and
contain one method that operates off of the ClipsEnvironment completely.

This allows me to work around an issue I had if I had use the Logical I/O
router system as the only point of entry for CLIPS to interact with the
rampancy framework.

This does have the added side effect that it is up to the compiler writer to
implement a CLIPS function to retrieve the proper information for compilation
or interpretation to occur.