fixed bugs dealing with compiling the rampancy library
I renamed and implemented the getters and setters for Context in
the Compiler class from get/setLLVMContext to get/setContext

I changed the CompilerRegistry to not store constant versions of
the Compiler objects. This is because I continually modify the pass as it's
being used for compilation and such.

I fixed some compilation errors in CompilerManager.cpp

I added DATA_OBJECT instances when compiling or interpreting anything in the
context of CLIPS and it turns out that the compiler does not exist. This was
necessary to make it possible to print messages out back to CLIPS.