Added the starting implementation of the Cortex class
I still need to implement convertToKnowledge. These functions generate a
PassManager and then add the appropriate Compiler and associated passes to
make it easy to perform knowledge construction.

The cortex also has a static instance that will be used by CLIPS to provide
durandal with an easy way to get access to the different compilation backends.