Added a static argv0 field to ClangCompiler.h
This is a hack/workaround the fact that we need to get the executable path as
part of the clang invocation process. This requires that the programmer set the
value at program startup. This value will persist across all ClangCompiler
instances.

This field is not used if you use compile(int argc, char** argv). This is
because it is then possible to get argv0 directly.