Simplified and cleaned up Cortex.cpp
I made it so that the call to convertToKnowledge(StringRef, Module) calls
convertToKnowledge(StringRef, Module, void*) where the void* is the environment
pointer of the environment saved within the Cortex instance.

I also fixed the function names passed to the CLIPS commands to be
rampancy-compile and rampancy-interpret instead of compile and interpret. This
is done mainly for completeness.