Heavily modified the clips dynamic compilation system
It no longer interfaces with a PassManager for execution and instead uses the
FunctionKnowledgeConversionPass which converts the different functions into
knowledge in the given module.

This change was made because I couldn't figure out why region and loop info
were not being allocated properly. I also didn't like that each compiler had to
be registered twice, once with the pass registry and another with the compiler
registry. I have changed it so that while the compiler registry is still used,
it does not double as a module pass for knowledge conversion. That task has
been given to the above pass and the Cortex.

I also setup the FunctionKnowledgeConvertionPass to be able to be set with a
target environment.

I also moved the argv0 command in ClangCompiler out into a separate class:
ClangCompilerGlobals for no real reason, I need to revert this change but for
   the time being the code does work.