Added loop simplify and break critical edges to the pass manager
This prevents an assertion error from occurring when getting unique exits from
a loop.

The break critical edges is something I automatically apply for wavefront
scheduling.

The code is however....broken as it attempts to cast an BinaryOperator as an
OverflowingBinaryOperator when it isn't. This is the next thing to look into.