Modified interpretToKnowledge and compileToKnowledge
compileToKnowledge no longer checks to make sure that the second argument is a
multifield. It only checks to make sure that there is at least a given compiler
name. It is up to the compiler backend to handle the rest of the arguments.

interpretToKnowledge has been modified so that the it takes in a symbol and
input string. So it checks for exactly two arguments.