Added support for having the contents of multiple functions existing in the environment
Originally, LibExpertSystem was designed with the assumption that each funciton
would be operated on individually within the expert system. The next function
would then be added after the current one was removed from the expert system.

The problem with this approach is that if I dynamically compile a file with
multiple functions then the names of the functions will overlap with one
another.

To fix this it is necessary to introduce ta new field to the FunctionNamer
class which is the name of the function that the FunctionNamer is referring to.
The name of this function is tacked onto the front of the name of each instance
contained in the expert system. Thus g0 becomes <fn-name>.g0 and so on.

I have updated the function namer to automatically do this if generic names are
requested. To solve the problem of prenamed llvm objects, it was necessary to
introduce concatFunctionNameToFront which will take the name of object and
concatenate the function name to the front of it. It has a form for a char*
buffer as well as a raw_string_ostream version as well.

The KnowledgeConstruction has been modified to take advantage of these new
features and no changes are necessary for any end-user code.

The other thing that the KnowledgeConstructor will do on function construction
is generate a fake name if it turns out that the given function does not have a
name. In this case, the name of the function will be "fn.<ptrtofunction>".