Fixed an issue with ConstantFP knowledge construction
It was possible for an assertion to be raised by LLVM because I was not
checking to see what kind of float semantics were defined for the given value.

This has been fixed so that if I find IEEEdouble or IEEEsingles then I cast the
value into CLIPS. Otherwise the value Unknown is set.