Added the start of the indirect clips library
This library is tasked with allowing clips to apply sets of passes (both
direct and indirect ones).

An indirect pass is one that will use LibIndirect to define a dynamic llvm pass
that acts as an entry point back into the expert system.

A direct pass is one that is written in C++ and compiled into a library or
module that is registered at program start.

The run-passes template defined in Templates.clp is one that can be used to
define a series of passes to run within an LLVM pass manager.