Added the Pass types
The pass types consist of two concrete classes (NativePass and Pass) which are
inherit from PassBase. The pass class represents an indirect LLVM pass while
the NativePass represents a direct LLVM pass.

The generation of the LLVM native passes is another thing all together. It will
requires the creation of a tool to make it easy to dynamically update the set
of passes based on different versions. Obviously, this tool does not exist yet.