Added headers for IndirectPasses and IndirectUniqueIdentifier
The IndirectPasses header contains abstract class definitions for all the
different instruction types. Some of the stranger pass types will not be fully
implemented yet but as time goes on this will change. The objective of these
abstract classes is to make it so that the required implementation is all
that's necessary. Normally, these classes are meant to provide template passes
that has execution logic which does not change.

IndirectUniqueIdentifier is a class/interface/mixin/protocol used to provide an
indirect way to uniquely identify a pass. It allocates a pointer to a single
char in memory. The pointer address is guarateed to be unique (unless you run
out of ram, at that point the program will crash regardless) which then can be
used to describe a unique Pass even though the class itself may be a template.