Added the header for IndirectPassGenerator
The IndirectPassGenerator is a way to generate base types for a given template
set. The base class IndirectPassGeneratorBase has a single method createPass
which returns an llvm::Pass* and takes in a IndirectPassHeader*. It is up to
subclasses to implement the rest of the functionality.

The subclass IndirectPassGenerator is a class that one defines the different
pass templates in. The rest of the functionality is handled automatically by
the class itself. One uses the PassTypeNotSupported class to mark a given pass
type as not supported.