Fixed compiler bugs in three different headers
I removed a cyclic issue within RegisterIndirectPass by making it so that the
implementation of the constructor has to be in a .cpp file.

I also removed all implementation details found in IndirectPassRegistry to
prevent compilation issues.

I modified the three boolean field in IndirectPassHeader to not be constant.
This was done because they become readonly after initialization.