overwrote the pointer fact in Environment
The only number allowed is zero which makes it easier to figure out if this is
a copy or not.

A message handler was added (init around) so that we can automatically
initialize an environment when a new environment object is created. It is up to
the end user to make the newly created environment aware of the creating
environment if they want.