Modified PrintCAddress in evaluatn.c to handle external addresses properly
PrintCAddress did not actually do a check to see if the call to
ValueToExternalAddress returned a proper value or not. This oversight does not
crash CLIPS but does cause the value <Pointer-C-(nil)> to be printed due to the
fact that the pointer is a null pointer.

I added a check to see if the result of ValueToExternalAddress was not a null
pointer. If it isn't then the original code is executed, otherwise gensprintf
is called but with theValue passed instead.