Modified the env commands to switch to the target env during execution
This fixes issues with commands that are inherent in clips and are not written
to take in a specific environment pointer. These commands use the
GetCurrentEnvironment command. Thus it is necesary to call
SetCurrentEnvironment(tgt) first, call the target function (even if it is an
		Env... function) and then call SetCurrentEnvironment(theEnv). This fixes
issues with commands such as batch*, etc, that use the current environment.

These changes are transparent to code within clips