Fixed an issue where getAnalysisUsage wasn't being called in indirect passes
It turns out that I forgot to have the const tag at the end of the function
declaration for each pass type's getAnalysisUsage call. This also required that
I tag applyIndirectAnalysisUsage to be const as well. This fixes the issue
where passes were not correctly firing because none of the analysis passes were
being marked as required.