Used the modify command instead of a retract-assert combo in MergeRules.clp
I realized that using the modify command allows me to retain the same memory
space while only changing the part that needs to be changed.

I also modified the settify rules to completely settify one of the target
multislots if a single duplicate is found. This should reduce the number of
expensive matches overall.