Added construction/WavefrontInitialization.clp
This file contains the rules originally found in the WavefrontInit stage.

This code has been rewritten so that the wavefront objects are created first.
This has the side effect of being able to figure out which regions are capable
of being wavefront scheduled because there is a wavefront object associated
with it.