Rewrote ValidBlockIdentification to be simpler and use modules
I found several areas where I was taking a fact with a multifield and
converting each element into it's own fact through the use of individual rule
fires. I believe at the time I was trying to be more "expert system" like but I
realized that it would be faster to use a progn$ in a single rule to do the
same thing. This increases the amount of work the rule does but reduces the
overall number of rules that need to be fired.