Added two new modules to Loader.clp
The two modules are wavefront-scheduling-identify and
wavefront-scheduling-pathing.

The first module (wavefront-scheduling-identify) is responsible for identifying
which blocks on the wavefront can be scheduled into. A block is considered a
valid target for scheduling into if it only has one successor. Blocks that have
more than one successor or no successors are not allowed.

The second module (wavefront-scheduling-pathing) is responsible for figuring
out the "range" of blocks that follow the target block which can be scheduled
out of and into the target block. The code used to identify this was rewritten
to not use the (return) command because that command removes an element from
the focus stack if modules are used. While the pipeline manager is robust, it
will be be able to continue execution if the focus stack has the pipeline
module removed from being executed next.

I also modified some of the internal facts generated by the modules to use
messages instead to allow me to use the modify command instead of retract and
assert. This reduces the amount of typing I have to do and, according to the
documentation, the memory that represents the original template is reused and
modified which saves space.