Started porting the rules in WavefrontScheduling.clp
I finished the wavefront-schedule-inject and wavefront-scheduling-strip
modules. I rewrote some of the more complex rules to copy block sets over if it
turns out that the multifield being stored into is empty. This makes some of
the rules constant in the best case and linear in the worst case now (this does
not include CLIPS' internal overhead though)