Added the post-ponder module
This module executes marks blocks as closed on the wavefront. This divergence
from the original code was necessary because the modification of next pass set
requires that we get back to the pipeline first. In the original code, the
changing of the substage fact would preclude execution of
FinishedSchedulingIntoBlock. However, in the module version it is necessary to
push FinishedSchedulingIntoBlock into its own module instead.