Fixed bugs in WavefrontPropagation.clp
I had given $?elements the name $?targets in my rewrite of one of the phi
functions. I have reverted this.

I had also made it so that the result of (/ (length$ $?elements) 2) was not
being bound to a variable and instead was being passed directly to the
create-phi-node function. This has been fixed.