added automatic initial-fact retract to the the pipeline
This is so that the initial facts asserted the pipeline do not clutter up the
set of facts. The fact is retracted after each module fires when control is
returned to the pipeline module. After the initial-fact(s) have occured then
the pipeline module will either go to the next pass or terminate if no passes
are left.

I also added a debug rule called DisplayFactsAtEnd. This is in the
wavefront-scheduling-final module and makes it really easy to find out if there
are dangling facts.