fixed an infinite loop bug in WavefrontPropagation.clp
This bug manfiests because each instance modification causes
PropagateAggregateInformation to restart its entire chain of execution. This
causes an infinite loop to occur.

I have reverted the body of the rule to be a slot-insert$ instead of modify
instance and have removed the slot match for InstuctionPropagation in the left
hand side of the rule.