Added the type name trait.
This trait is meant to provide a standardize way to defining the CLIPS class
name of a given C++ class. For instance, llvm::Value would be known as
LLVMValue in CLIPS. So this typename trait would tell the compiler that when
you see a llvm::Value that it's CLIPS type name is LLVMValue.