Added the header for the rewritten CLIPSObjectBuilder class
This new class is known as KnowledgeRepresentationBuilder because nothing
inherits from it in the new design. This class is still responsible for
building an instance string. Some of the functions have been renamed to be more
meaningful as well. For instance, openField and closeField are now known as
openMultifield and closeMultifield.

I also change the data fields to be pointers to data items to allow the
KnowledgeRepresentationBuilder to exist separately of a routing chain. While
there is no real reason for this, it does make my life easier when debugging.