Added custom name generation logic for basic blocks and functions
Providing this logic makes it so that the name of a given object can be
retrieved without routing the object. The reasons for this may be that the
object that would call the routing doesn't have enough information to correct
construct the knowledge representation of the given object, or the target
object doesn't actually need to be represented in CLIPS beyond it's name.