Added a new void* field to the KnowledgeConstructor class
This field is called targetObject and is meant to store a pointer to the object
being targeted by the current pass. This translates to the object passed to the
pass to be operated on. It is a void* so that the addresses can be compared
because that is what is important.